Add classes for each path
