+++ src/main/java/frc/robot/autonomous/Path2n21gBlue.java
@@ -1,0 +1,28 @@
+package frc.robot.autonomous;
+
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.commandGroups.MiddleScoring;
+import frc.robot.commandGroups.UpperScoring;
+import frc.robot.subsystems.drivetrain.SwerveDrive;
+import frc.robot.subsystems.drivetrain.commands.BalanceOnStation;
+
+public class Path2n21gBlue extends AutonomousPaths{
+    private final SwerveDrive swerveDrive = SwerveDrive.getInstance();
+
+    protected CommandBase placeUpperCone21AndCharge() {
+        return new SequentialCommandGroup(
+                new UpperScoring(),
+                FollowPath.loadTrajectory(".pathplanne/2n21g blue", swerveDrive),
+                new BalanceOnStation()
+        );
+    }
+
+    protected CommandBase placeMiddleCone21AndCharge() {
+        return new SequentialCommandGroup(
+                new MiddleScoring(),
+                FollowPath.loadTrajectory(".pathplanne/2n21g blue", swerveDrive),
+                new BalanceOnStation()
+        );
+    }
+}
